ChatBot/requirements.txt
streamlit
google-generativeai
python-dotenv
pathway[xpack-llm,xpack-llm-docs]
langchain
langchain-community
langchain-text-splitters
openai
docling
PyPDF2